requirements.txt
celery==5.2.3
Django==4.0.2
django-mptt==0.13.4
django-mptt-admin==2.2.0
Pillow==9.0.1
psycopg2-binary==2.9.3
redis==4.1.4
requests==2.27.1
vk-api==11.9.7
python-dotenv==0.20.0
yadisk==1.2.16